reservations: Backport optimization and newer system of budget tracking
A reservation with background priority is always lowest priority
and can always be added to the end of the sorted reservations list,
so don't bother to search the list in those cases.

The newer system of budget tracking does not charge ACTIVE_IDLE
reservations. This is the desired approach for MC^2. Otherwise,
if a Level-B client entered while a Level-A client was still
executing, the Level-B client would not recieve its full share
of budget. (Recall that ACTIVE_IDLE reservations are those with
budget but no clients.)