[MC^2 Scheduler] Import core MC^2 scheduler components
This comes from the wip-kernel-coloring branch, commit
cfa114e7818006b0fea6914294c2ef9fc4bcbbb8. Only change on
top of the import should be the renumbering of MC^2's
trace IDs in `trace.h` to avoid conflicting with those
added in recent versions of LITMUS^RT.

NOTE: This WILL NOT BUILD, and is being committed to enable
historical tracking. The follow-up commits fix these build
issues, but this commit is needed to make it clear what
was changed after importing from the legacy version of
LITMUS^RT.