[MC^2 Scheduler] Re-implement Level-C scheduling with extended reservations
Extended reservations compartmentalize Level-C scheduling, making
it easier to swap out Level-C schedulers or implement clustering.

As of this commit, collections of Level-A, -B, and -C tasks seem
to run without issue with the important caveat that I don't believe
Level-A tasks are correctly prioritized over Level-B (this has been
a standing problem).