Unify the MC^2 API with LITMUS's and use separate L-A and -B reservation envs
Finally removes mc2_task by combining task_params.cpu with res_id as
in P-RES and by using task_params.cls for the criticality level.

Also splits Level-A and Level-B into separate reservation
environments so that priority inversions are not possible. As part
of this change, only Level-A reservations can be explicitly created.
If not explicitly created, the MC^2 scheduler will implicitly create
a reservation for the task.

Contains major userspace API changes and is NOT compatible with old
MC^2 programs, but is compatible with traditional LITMUS-RT
programs such as rtspin.

Also includes some comment cleanup in sched_mc2.c