GSN-EDF: Make sleep_next_period() a no-op if already completed due to budget exhaustion
Vanilla LITMUS-RT (without reservations) does not distinguish between
a period which voluntarily ends (via the complete_job() userspace call)
and a period which is forced to complete due to budget exhaustion.

This results in inconsistent behavior of the userspace API when budget
enforcement is enabled. Say a task's budget is set slightly below its actual
execution time. In this case, the task is suspended during its first job by
budget enforcement. When the second period starts, the task immediately
completes it's work and calls sleep_next_period(). From the perspective of
the task, it should enter its second job after calling sleep_next_period().
From the perspective of LITMUS-RT, the task is already in its second
period and will suspend the task until its _third_ period. This patch fixes
that bug for GSN-EDF by only suspending the task on
sleep_next_period() if the previous period was not force-completed by
budget enforcement.